requirements.txt
telethon>=1.40.0
python-telegram-bot>=22.1
openai>=1.0.0
httpx>=0.28.0
qrcode>=8.0
python-dotenv>=1.0.1